camoufox-py/requirements.txt
PyYAML==6.0.2
camoufox[geoip]==0.4.11